requirements.txt
certifi==2019.3.9
chardet==3.0.4
docutils==0.14
idna==2.8
Kivy==1.10.1
Kivy-examples==1.10.1
Kivy-Garden==0.1.4
kivy.deps.angle==0.1.7
kivy.deps.glew==0.1.10
kivy.deps.gstreamer==0.1.13
kivy.deps.sdl2==0.1.18
numpy==1.16.3
opencv-contrib-python==4.1.0.25
opencv-python==4.1.0.25
Pillow==6.0.0
Pygments==2.3.1
pypiwin32==223
pywin32==224
requests==2.21.0
urllib3==1.24.2
